Updated README.md and updated interface with base abstract class and pyproject.toml for configuration
