+++ ai_conversation_client.egg-info/PKG-INFO
@@ -1,0 +1,262 @@
+Metadata-Version: 2.4
+Name: ai-conversation-client
+Version: 0.1.0
+Summary: A Python client for interacting with AI language models
+Author-email: Your Name <your.email@example.com>
+License: MIT
+Classifier: Programming Language :: Python :: 3
+Classifier: License :: OSI Approved :: MIT License
+Classifier: Operating System :: OS Independent
+Requires-Python: >=3.8
+Description-Content-Type: text/markdown
+License-File: LICENSE
+Requires-Dist: openai>=1.0.0
+Requires-Dist: python-dotenv>=1.0.0
+Provides-Extra: dev
+Requires-Dist: pytest>=7.4.0; extra == "dev"
+Requires-Dist: pytest-cov>=4.1.0; extra == "dev"
+Requires-Dist: black>=23.7.0; extra == "dev"
+Requires-Dist: ruff>=0.0.285; extra == "dev"
+Requires-Dist: mypy>=1.5.1; extra == "dev"
+Dynamic: license-file
+
+# C++ Template Repository
+
+Welcome to the **C++ Template Repository**! This project provides a quick-start foundation for developing C++ applications with:
+
+- **CMake** for build management
+- **vcpkg** for dependency management (e.g., GoogleTest)
+- **clang-format** and **clang-tidy** for style and static analysis
+- **CircleCI** for continuous integration (build, test, coverage)
+- **LLVM** coverage tools (`llvm-profdata`, `llvm-cov`) or `lcov/gcov` (configurable)
+- **Issue** and **Pull Request** templates to standardize contributions
+
+---
+
+## Features
+
+- **Modern CMake** project setup (`CMakeLists.txt` at the root)
+- **Dependency management** with vcpkg for easy library installations (e.g., GoogleTest)
+- **Unit, integration, and end-to-end tests** using GoogleTest
+- **clang-format** for code formatting; **clang-tidy** for static analysis
+- **Coverage** reports generated by `llvm-cov`/`llvm-profdata` or `lcov/gcov`
+- **CircleCI** configuration that:
+  - Installs dependencies (Clang, CMake, vcpkg, etc.)
+  - Builds and tests the code
+  - Collects coverage data
+  - Uploads coverage reports as artifacts
+- **Preconfigured** Issue and Pull Request templates for standardized contributions
+
+---
+
+## CircleCI Status
+
+[![CircleCI](https://dl.circleci.com/status-badge/img/gh/amoghkr1shna/team2/tree/hw1.svg?style=shield)](https://dl.circleci.com/status-badge/redirect/gh/amoghkr1shna/team2/tree/hw1)
+
+---
+
+## Prerequisites
+
+### Local Development
+
+- **C++ Compiler** (Clang or GCC)
+- **CMake** (3.15+ recommended)
+- **vcpkg** (optional but recommended for cross-platform dependency management)
+- **clang-format** and **clang-tidy** (for style and static analysis)
+- **GoogleTest** (installed via vcpkg or your system package manager)
+
+##### macOS (Homebrew example)
+
+```bash
+# Install CMake, LLVM, and GoogleTest
+brew install cmake llvm googletest
+
+# Or install vcpkg for cross-platform dependencies
+git clone https://github.com/microsoft/vcpkg.git
+cd vcpkg
+./bootstrap-vcpkg.sh
+./vcpkg integrate install
+./vcpkg install gtest
+```
+
+##### Ubuntu (apt-get example)
+
+```bash
+sudo apt-get update
+sudo apt-get install -y clang cmake build-essential lcov
+# For vcpkg:
+git clone https://github.com/microsoft/vcpkg.git
+cd vcpkg
+./bootstrap-vcpkg.sh
+./vcpkg integrate install
+./vcpkg install gtest
+```
+
+#### Continuous Integration
+
+A CircleCI account connected to your GitHub (or GitLab) repository.
+(Optional) A GitHub repository with a .circleci/config.yml file in the root.
+
+---
+
+## Getting Started
+
+### Clone this repository:
+
+```bash
+git clone https://github.com/<your-username>/<cpp-template-repo>.git
+cd <cpp-template-repo>
+```
+
+### Install Dependencies:
+
+If using vcpkg, install required packages (e.g., gtest):
+
+```bash
+./vcpkg/vcpkg install gtest
+```
+
+Ensure clang-format and clang-tidy are installed for local checks.
+(Optional) Configure vcpkg:
+
+```bash
+cmake -B build -DCMAKE_TOOLCHAIN_FILE=./vcpkg/scripts/buildsystems/vcpkg.cmake
+```
+
+---
+
+## Building
+
+Configure and build the project using CMake:
+
+```bash
+mkdir build && cd build
+cmake -DCMAKE_CXX_FLAGS="--coverage" \
+ -DCMAKE_TOOLCHAIN_FILE=../vcpkg/scripts/buildsystems/vcpkg.cmake \
+ ..
+make -j4
+```
+
+The `--coverage` flag is optional and adds instrumentation for coverage analysis.
+Adjust paths if your vcpkg folder is elsewhere or if you are not using vcpkg.
+
+---
+
+## Testing
+
+Tests are organized as:
+
+Unit tests (e.g., unit_tests executable)
+Integration tests (e.g., integration_tests)
+End-to-End tests (e.g., e2e_tests)
+
+---
+
+## Coverage
+
+This repository supports coverage using either LLVM or lcov/gcov.
+
+### LLVM Coverage Example
+
+Build with coverage flags.
+Run tests to generate .profraw files:
+
+```bash
+export LLVM_PROFILE_FILE="%p.profraw"
+ctest --output-on-failure
+```
+
+### Collect coverage:
+
+```bash
+lcov --capture --directory . --output-file coverage.info
+lcov --remove coverage.info '/usr/_' 'tests/_' --output-file coverage.info
+genhtml coverage.info --output-directory coverage_html
+open coverage_html/index.html
+```
+
+---
+
+## Style & Static Analysis
+
+### clang-format
+
+A .clang-format file is included. To auto-format your code:
+
+```bash
+clang-format -i src/_.cpp src/_.hpp tests/\*_/_.cpp
+```
+
+In CI, a dry-run with --Werror ensures that code must adhere to the formatting rules.
+
+### clang-tidy
+
+A .clang-tidy file is provided, which includes checks like modernize-_ and readability-_. Run clang-tidy as follows:
+
+```bash
+clang-tidy -p build <file> -- -std=c++17
+```
+
+If certain rules are too strict (e.g., trailing return types or short parameter names), adjust the checks in the .clang-tidy file or fix your code accordingly.
+
+---
+
+## Continuous Integration (CircleCI)
+
+The repository includes a CircleCI configuration (.circleci/config.yml) that:
+
+- Installs dependencies (Clang, CMake, vcpkg, etc.)
+- Bootstraps vcpkg and installs GoogleTest
+- Builds the project with coverage flags
+- Runs tests (producing JUnit XML output for the CircleCI Tests tab)
+- Generates a coverage report (HTML)
+- Uploads coverage artifacts for easy access
+- Viewing CI Results
+- Push your code to GitHub (or create a pull request).
+- CircleCI will automatically trigger a build.
+- Checks the Tests tab for test results.
+- Checks the Artifacts tab for the coverage report (in the coverage_html folder).
+
+---
+
+## Issue & PR Templates
+
+This repository contains preconfigured templates for issues and pull requests:
+
+```
+.github/ISSUE_TEMPLATE/bug_report.md
+.github/ISSUE_TEMPLATE/feature_request.md
+pull_request_template.md (in the root or in .github/PULL_REQUEST_TEMPLATE/)
+```
+
+These templates ensure that all contributions follow a consistent format.
+
+---
+
+## Contributing
+
+Contributions are welcome! To contribute:
+
+1. Fork this repository
+2. Create a new branch (e.g., feature/your-feature):
+   ```bash
+   git checkout -b feature/your-feature
+   ```
+3. Implement your changes (ensure code is formatted and linted):
+   ```bash
+   git commit -m "Add your feat."
+   ```
+4. Push to the branch:
+   ```bash
+   git push origin feature/your-feature
+   ```
+5. Open a Pull Request and follow the provided template.
+6. Ensure all CircleCI checks (formatting, static analysis, tests, and coverage) pass.
+
+Feel free to open an issue for any feature requests or bug reports.
+
+---
+
+## License
+
+This project is licensed under the MIT License. See the [LICENSE](LICENSE) file for details.
