+++ ai_conversation_client.egg-info/requires.txt
@@ -1,0 +1,9 @@
+openai>=1.0.0
+python-dotenv>=1.0.0
+
+[dev]
+pytest>=7.4.0
+pytest-cov>=4.1.0
+black>=23.7.0
+ruff>=0.0.285
+mypy>=1.5.1
