+++ ai_conversation_client/__init__.py
@@ -1,15 +1,16 @@
 """
 AI Conversation Client
 
-A Python client for interacting with OpenAI's ChatGPT API.
+A Python client for interacting with AI language models.
 This module provides an interface for sending messages, managing conversations,
-and handling responses from the OpenAI API.
+and handling responses from various AI providers.
 
 Usage:
-    from ai_conversation_client import AIClient, Message, MessageRole
+    from ai_conversation_client import BaseAIClient, Message, MessageRole
+    from ai_conversation_client.providers import OpenAIClient
     
-    # Create a client
-    client = AIClient(api_key="your_api_key")
+    # Create a client with your preferred provider
+    client = OpenAIClient(api_key="your_api_key")
     
     # Create a conversation
     conversation = client.create_conversation(title="My Chat", 
@@ -27,8 +28,8 @@ import uuid
 import os
 import json
 import asyncio
+from abc import ABC, abstractmethod
 from dotenv import load_dotenv
-from openai import AsyncOpenAI
 
 # Load environment variables
 load_dotenv()
@@ -215,34 +216,12 @@ class Conversation:
         return conversation
 
 
-class AIClient:
-    """Client for interacting with OpenAI's ChatGPT API."""
+class BaseAIClient(ABC):
+    """Abstract base class for AI conversation clients."""
     
-    def __init__(self, api_key: Optional[str] = None, 
-                 model: str = "gpt-3.5-turbo", 
-                 temperature: float = 0.7,
-                 max_tokens: Optional[int] = None):
-        """
-        Initialize a new AI client.
-        
-        Args:
-            api_key: OpenAI API key (can be None if set in environment)
-            model: The OpenAI model to use (default: gpt-3.5-turbo)
-            temperature: Controls randomness (0-1, default: 0.7)
-            max_tokens: Maximum tokens in response (None for model default)
-        """
-        self._api_key = api_key or os.getenv("OPENAI_API_KEY")
-        if not self._api_key:
-            raise ValueError("OpenAI API key is required. Provide it as a parameter or set it in the OPENAI_API_KEY environment variable.")
-        
-        self._model = model or os.getenv("MODEL", "gpt-3.5-turbo")
-        self._temperature = temperature if temperature is not None else float(os.getenv("TEMPERATURE", "0.7"))
-        self._max_tokens = max_tokens or (int(os.getenv("MAX_TOKENS")) if os.getenv("MAX_TOKENS") else None)
+    def __init__(self):
+        """Initialize a new AI client."""
         self._conversations: Dict[str, Conversation] = {}
-        
-        # Initialize the OpenAI client
-        # Use simpler initialization to avoid compatibility issues
-        self._client = AsyncOpenAI(api_key=self._api_key)
     
     def create_conversation(self, title: Optional[str] = None, 
                           system_prompt: Optional[str] = None) -> Conversation:
@@ -286,6 +265,7 @@ class AIClient:
         """
         return list(self._conversations.values())
     
+    @abstractmethod
     async def send_message(self, conversation_id: str, message_content: str) -> Message:
         """
         Send a user message to the AI and get a response.
@@ -302,58 +282,26 @@ class AIClient:
             
         Raises:
             ValueError: If the conversation ID is not found
-            Exception: If there is an error communicating with the OpenAI API
+            Exception: If there is an error communicating with the AI provider
             
         Usage:
             response = await client.send_message(conversation.id, "Hello, AI!")
             print(response.content)
         """
-        conversation = self.get_conversation(conversation_id)
-        if not conversation:
-            raise ValueError(f"Conversation with ID {conversation_id} not found")
-        
-        # Add user message to conversation
-        user_message = Message(message_content, MessageRole.USER)
-        conversation.add_message(user_message)
-        
-        # Prepare messages for the API request
-        messages = [msg.to_dict() for msg in conversation.messages]
-        
-        try:
-            # Call the OpenAI API
-            response = await self._client.chat.completions.create(
-                model=self._model,
-                messages=messages,
-                temperature=self._temperature,
-                max_tokens=self._max_tokens
-            )
-            
-            # Extract the response content
-            response_content = response.choices[0].message.content
-            
-            # Create a message from the response
-            ai_message = Message(response_content, MessageRole.ASSISTANT)
-            
-            # Add AI response to conversation
-            conversation.add_message(ai_message)
-            
-            return ai_message
-        except Exception as e:
-            # Handle API errors
-            error_message = f"Error communicating with OpenAI API: {str(e)}"
-            raise ConnectionError(error_message) from e
+        pass
     
+    @abstractmethod
     def set_model(self, model: str) -> None:
         """
         Change the AI model used for generating responses.
         
         Args:
-            model: The name of the OpenAI model to use (e.g., "gpt-3.5-turbo", "gpt-4")
+            model: The name of the model to use
             
         Usage:
             client.set_model("gpt-4")
         """
-        self._model = model
+        pass
     
     def export_conversation(self, conversation_id: str, 
                           format: str = "json") -> Union[str, Dict[str, Any]]:
