+++ ai_conversation_client/providers/__init__.py
@@ -1,0 +1,10 @@
+"""
+AI Conversation Client Providers
+
+This package contains implementations of the BaseAIClient interface
+for different AI providers.
+"""
+
+from ai_conversation_client.providers.openai import OpenAIClient
+
+__all__ = ["OpenAIClient"]
